# docs/swagger.yaml
basePath: /api
definitions:
  auth.GuestLoginRequest:
    properties:
      guest_id:
        type: string
    type: object
  auth.GuestLoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.HelloResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      google_id:
        type: string
      guest_id:
        type: string
      id:
        type: string
      is_guest:
        type: boolean
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  models.UserRole:
    enum:
    - USER
    - ADMIN
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
  websocket.CreateRoomRequest:
    properties:
      max_players:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  websocket.InviteRequest:
    properties:
      room_id:
        type: string
      user_ids:
        items:
          type: string
        type: array
    required:
    - room_id
    - user_ids
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Go starter example project
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Starter Example ProjectAPI
  version: 1.0.0
paths:
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Returns admin dashboard information - only accessible by admins
      produces:
      - application/json
      responses:
        "200":
          description: Admin dashboard data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get admin dashboard data
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Returns a list of all users in the system - only accessible by
        admins
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /auth/guest-login:
    post:
      consumes:
      - application/json
      description: Creates a new guest user or logs in existing guest with guest_id
      parameters:
      - description: Guest login request (optional guest_id)
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.GuestLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.GuestLoginResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Guest login
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Returns the current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Validates refresh token only and issues new tokens. Old refresh
        token will be blacklisted.
      parameters:
      - description: Refresh token only
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.GuestLoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh token
      tags:
      - auth
  /hello:
    get:
      consumes:
      - application/json
      description: Returns a hello message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HelloResponse'
      summary: Hello endpoint
      tags:
      - general
  /ws:
    get:
      description: Establish WebSocket connection for real-time communication
      parameters:
      - description: Room ID to join (use 'lobby' for public lobby)
        in: query
        name: room_id
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Connect to WebSocket
      tags:
      - websocket
  /ws/invite:
    post:
      consumes:
      - application/json
      description: Invite specific users to a game room (admin only)
      parameters:
      - description: Invitation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/websocket.InviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitations sent successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Invite users to room
      tags:
      - websocket
  /ws/rooms:
    get:
      description: Get list of all active WebSocket rooms
      responses:
        "200":
          description: List of rooms
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all rooms
      tags:
      - websocket
    post:
      consumes:
      - application/json
      description: Create a new game room (admin only)
      parameters:
      - description: Room details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/websocket.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Room created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create game room
      tags:
      - websocket
  /ws/rooms/{room_id}:
    delete:
      description: Close an existing game room (admin only)
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: string
      responses:
        "200":
          description: Room closed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Cannot close lobby
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Close game room
      tags:
      - websocket
    get:
      description: Get detailed information about a specific room
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: string
      responses:
        "200":
          description: Room information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Room not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get room information
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"